fixed scale bar unit painting
git-svn-id: file:///home/svn/shoola-omero/trunk@4040 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/agents/imviewer/browser/BrowserCanvas.java
@@ -167,7 +167,7 @@ class BrowserCanvas
         int size = (int) model.getUnitBarSize();
         if (v > 0 && model.isUnitBar()) {
             paintScaleBar(g2D, img.getWidth()-size-5, 
-                        img.getHeight()-10, size, ""+(int) v);
+                        img.getHeight()-10, size, ""+v);
         }
     }
